R2T/requirements.txt
django
djangorestframework
markdown  
django-filter   
stripe
googletrans
serializers
channels==3.0.4
gunicorn==20.1.0
daphne==3.0.2